[REFACTOR]: HttpStatus, ResponseEnum 위치 변경
