+++ src/main/java/model/HttpStatus.java
@@ -1,4 +1,4 @@
-package webserver;
+package model;
 
 public enum HttpStatus {
     OK("200", "OK"),
