+++ src/test/java/webserver/ResponseEnumTest.java
@@ -1,5 +1,6 @@
 package webserver;
 
+import model.ResponseEnum;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
